requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0
mysql-connector-python>=8.0.0

